CAN Interface initialization is done in ICanHelper interface instead ICanSender interface.
It was not possible to initialize more than one CanSender and therefore,
use different senders. The same will happen when trying to initialize
different CanReceivers (future implementation) as they will try to
initialize the interfaces when they are already initialized.

Therefore, the initialization must be centralized by the ICanHelper
interface implementation so that the interfaces are initialized once by
ICanHelper at the beggining of the app execution instead of several
times by CanSenders/CanReceivers.

CanSenders/Receivers will initialize itself locally (take resources,
etc), but the global initialization of the interface will be done by
CanHelper (for example, setting the interface up for SocketCan or
initializing the interface for PeakCan).